refactor(glm5): split stream parser responsibilities
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
